+++ Memberships/Memberships/Areas/Admin/Controllers/PartController.cs
@@ -112,8 +112,13 @@ namespace Memberships.Areas.Admin.Controllers
         public async Task<ActionResult> DeleteConfirmed(int id)
         {
             Part part = await db.Parts.FindAsync(id);
-            db.Parts.Remove(part);
-            await db.SaveChangesAsync();
+            var isUnused = await db.Items.CountAsync(i => i.PartId.Equals(id)) == 0;
+
+            if (isUnused)
+            {
+                db.Parts.Remove(part);
+                await db.SaveChangesAsync();
+            }
             return RedirectToAction("Index");
         }
