+++ Memberships/Memberships/Areas/Admin/Controllers/ProductTypeController.cs
@@ -112,8 +112,15 @@ namespace Memberships.Areas.Admin.Controllers
         public async Task<ActionResult> DeleteConfirmed(int id)
         {
             ProductType productType = await db.ProductTypes.FindAsync(id);
-            db.ProductTypes.Remove(productType);
-            await db.SaveChangesAsync();
+
+            var isUnused = await db.Products.CountAsync(i => i.ProductTypeId.Equals(id)) == 0;
+
+            if (isUnused)
+            {
+                db.ProductTypes.Remove(productType);
+                await db.SaveChangesAsync();
+            }
+
             return RedirectToAction("Index");
         }
